CoreFAPI/BackApp/groups_interest_service/290325/f/interest_groups/requirements.txt
# Основные зависимости (без жестких ограничений версий)
fastapi[all]>=0.95.0  # [all] включает дополнительные возможности
uvicorn[standard]>=0.20.0  # С дополнительными компонентами
python-multipart>=0.0.6
pydantic-settings>=2.0.0
fastapi-cache2>=0.2.0

# Базы данных (используем последние версии)
sqlalchemy[asyncio]>=2.0.0  # Включает async поддержку
psycopg2-binary>=2.9.6
redis[async]>=4.5.5  # Включает aioredis
asyncpg>=0.27.0
alembic>=1.11.1
python-dateutil>=2.8.2  # Для работы с датами

# Аутентификация
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
cryptography>=40.0.0

# RabbitMQ (последние стабильные)
aio-pika>=9.0.7
aiormq>=6.7.0

# HTTP клиенты
httpx>=0.25.0
requests>=2.31.0

# Логирование (удобные для разработки)
loguru>=0.7.0  # Очень удобный логгер для dev
rich>=13.0.0  # Красивое форматирование вывода

# Dev-зависимости (добавлены полезные инструменты)
pytest>=7.3.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-xdist>=3.0.0  # Параллельное тестирование
pytest-httpx>=0.27.0
responses>=0.24.1
ipython>=8.0.0  # Улучшенный REPL
watchfiles>=0.19.0  # Автоперезагрузка
httpx-sse>=0.4.0
factory-boy>=3.2.0
faker>=18.11.1
mypy>=1.3.0
types-requests>=2.31.0
types-redis>=4.5.5
debugpy>=1.6.0  # Отладка в VS Code